Added Enemy and Minor Changes
Created Skeleton scene and a lot of scripts based on the original Player scripts, including the State Machine. Added a sprite for the Skelton Enemy and stripped out the player controls. Also made minor edits to the file structure of the State Machines and made the Player Animations easier to follow/edit.